restructure pipeline directory and refactor import statements where relevant
