+++ asf_heat_pump_suitability/pipeline/enhance_epc/enhance_epc.py
@@ -1,6 +1,6 @@
 import polars as pl
 from asf_heat_pump_suitability.getters import get_datasets
-from asf_heat_pump_suitability.pipeline import prepare_features
+from asf_heat_pump_suitability.pipeline.prepare_features import output_areas
 
 
 def join_df_additional_features(epc_df: pl.DataFrame) -> pl.DataFrame:
@@ -13,8 +13,8 @@ def join_df_additional_features(epc_df: pl.DataFrame) -> pl.DataFrame:
     Returns
         pl.DataFrame: EPC dataset with additional feature columns
     """
-    epc_df = prepare_features.standardise_col_postcode(epc_df, pcd_col="POSTCODE")
-    onspd_df = prepare_features.prepare_df_ons_pd()
+    epc_df = output_areas.standardise_col_postcode(epc_df, pcd_col="POSTCODE")
+    onspd_df = output_areas.prepare_df_ons_pd()
 
     df = epc_df.join(onspd_df, how="left", on="POSTCODE")
