+++ asf_heat_pump_suitability/pipeline/prepare_features/output_areas.py
